Undo my lib-binary split
The idea was to do special stuff in a separate crate in order to have
starboard run as your desktop on X11, but as I've dove deeper into that,
I think it's a bad idea to handle that this way.

Using your wm or dedicating a tag/workspace for starboard is probably a
better idea.